refactor(api): move security-pre-handler usecases into shared
