+++ api/tests/shared/integration/application/usecases/checkAuthorizationToAccessCampaign_test.js
@@ -1,6 +1,6 @@
-import * as checkAuthorizationToAccessCampaign from '../../../../src/shared/application/usecases/checkAuthorizationToAccessCampaign.js';
-import { expect } from '../../../test-helper.js';
-import { databaseBuilder } from '../../../tooling/databases.js';
+import * as checkAuthorizationToAccessCampaign from '../../../../../src/shared/application/usecases/checkAuthorizationToAccessCampaign.js';
+import { expect } from '../../../../test-helper.js';
+import { databaseBuilder } from '../../../../tooling/databases.js';
 
 describe('Integration | API | checkAuthorizationToAccessCampaign', function () {
   describe('when the user belongs to organization', function () {
