+++ api/tests/shared/integration/application/usecases/checkUserIsAdminAndManagingStudentsForOrganization_test.js
@@ -1,7 +1,7 @@
-import * as checkUserIsAdminAndManagingStudentsForOrganization from '../../../../src/shared/application/usecases/checkUserIsAdminAndManagingStudentsForOrganization.js';
-import { Membership } from '../../../../src/shared/domain/models/Membership.js';
-import { expect } from '../../../test-helper.js';
-import { databaseBuilder } from '../../../tooling/databases.js';
+import * as checkUserIsAdminAndManagingStudentsForOrganization from '../../../../../src/shared/application/usecases/checkUserIsAdminAndManagingStudentsForOrganization.js';
+import { Membership } from '../../../../../src/shared/domain/models/Membership.js';
+import { expect } from '../../../../test-helper.js';
+import { databaseBuilder } from '../../../../tooling/databases.js';
 
 describe('Integration | API | checkUserIsAdminAndManagingStudentsForOrganization', function () {
   describe('when the user does not belongs to the organization', function () {
